pyproject.toml
[project]
name = "agentic-internet-rsearch-deepagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "deepagents>=0.0.5",
    "langchain-anthropic>=0.3.20",
    "langchain-community>=0.3.29",
    "langchain-mcp-adapters>=0.1.10",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
]
